move local_clickzetta_settings to sqlglot module
