+++ sqlglot/local_clickzetta_settings.py
